+++ clean_python/fastapi/security.py
@@ -1,80 +1,90 @@
-from fastapi import Depends
-from fastapi import Request
-from fastapi.security import HTTPBearer
-from fastapi.security import OAuth2AuthorizationCodeBearer
+from typing import Callable
 
+from fastapi import Request
+from fastapi.security import OAuth2AuthorizationCodeBearer
+from fastapi.security import SecurityScopes
+
+from clean_python import ctx
 from clean_python import PermissionDenied
-from clean_python.oauth2 import BaseTokenVerifier
-from clean_python.oauth2 import NoAuthTokenVerifier
-from clean_python.oauth2 import OAuth2SPAClientSettings
+from clean_python import Scope
+from clean_python import ValueObject
+from clean_python.oauth2 import OAuth2Settings
 from clean_python.oauth2 import Token
 from clean_python.oauth2 import TokenVerifier
 from clean_python.oauth2 import TokenVerifierSettings
 
-__all__ = ["get_token", "RequiresScope"]
+__all__ = ["get_token", "default_scope_verifier", "AuthSettings"]
 
-verifier: BaseTokenVerifier | None = None
-
-
-def clear_verifier() -> None:
-    global verifier
-
-    verifier = None
+ScopeVerifier = Callable[[Request, Scope, Token], None]
 
 
-def set_verifier(settings: TokenVerifierSettings | None) -> None:
-    global verifier
+class OAuth2Schema(OAuth2AuthorizationCodeBearer):
+    """A fastapi 'dependable' to verify bearer tokens obtained through OAuth2.
 
-    if settings is None:
-        verifier = NoAuthTokenVerifier()
-    else:
-        verifier = TokenVerifier(settings=settings)
+    This verification includes authentication and scope verification.
 
-
-def get_token(request: Request) -> Token:
-    """A fastapi 'dependable' yielding the validated token"""
-    global verifier
-
-    assert verifier is not None
-    return verifier(request.headers.get("Authorization"))
-
-
-class RequiresScope:
-    def __init__(self, scope: str):
-        assert scope.replace(" ", "") == scope, "spaces are not allowed in a scope"
-        self.scope = scope
-
-    async def __call__(self, token: Token = Depends(get_token)) -> None:
-        if self.scope not in token.scope:
-            raise PermissionDenied(f"this operation requires '{self.scope}' scope")
-
-
-class OAuth2SPAClientSchema(OAuth2AuthorizationCodeBearer):
-    """A fastapi 'dependable' configuring the openapi schema for the
-    OAuth2 Authorization Code Flow with PKCE extension.
-
-    This includes the JWT Bearer token configuration.
+    Because this class derives from a FastAPI built in, the openapi schema for the
+    OAuth2 Authorization Code Flow is automatically configured.
     """
 
-    def __init__(self, client: OAuth2SPAClientSettings):
+    def __init__(self, settings: "AuthSettings"):
+        self._verifier = TokenVerifier(settings.token)
+        self._scope_verifier = settings.scope_verifier
         super().__init__(
-            scheme_name="OAuth2Bearer",
-            authorizationUrl=str(client.authorization_url),
-            tokenUrl=str(client.token_url),
+            scheme_name="OAuth2",
+            authorizationUrl=settings.oauth2.authorization_url,
+            tokenUrl=settings.oauth2.token_url,
+            scopes=settings.oauth2.scopes,
         )
 
-    async def __call__(self) -> None:
-        pass
+    async def __call__(
+        self, request: Request, security_scopes: SecurityScopes
+    ) -> Token:
+        """FastAPI magically fills these parameters:
+
+        - 'request' with the current request
+        - 'security_scopes' with the union of all Security(..., scopes) in use for
+          the current operation. Note that this also automatically configures the
+          "security requirements" in the openapi spec for this operation.
+        """
+        token = self._verifier(request.headers.get("Authorization"))
+        self._scope_verifier(request, frozenset(security_scopes.scopes), token)
+        ctx.user = token.user
+        ctx.tenant = token.tenant
+        return token
 
 
-class JWTBearerTokenSchema(HTTPBearer):
-    """A fastapi 'dependable' configuring the openapi schema for JWT Bearer tokens.
+# the scheme is stored globally enabling for the "get_token" callable
+scheme: OAuth2Schema | None = None
 
-    Note: for the client-side OAuth2 flow, use OAuth2SPAClientSchema instead.
-    """
 
-    def __init__(self):
-        super().__init__(scheme_name="OAuth2Bearer", bearerFormat="JWT")
+async def get_token(request: Request) -> Token:
+    """A fastapi 'dependable' yielding the validated token"""
+    global scheme
+    assert scheme is not None
+    return await scheme(request, SecurityScopes())
 
-    async def __call__(self) -> None:
-        pass
+
+def default_scope_verifier(
+    request: Request, endpoint_scopes: Scope, token: Token
+) -> None:
+    """Verifies whether any of the endpoint_scopes is in the token."""
+    if not all(x in token.scope for x in endpoint_scopes):
+        raise PermissionDenied(
+            f"this operation requires '{' '.join(endpoint_scopes)}' scope"
+        )
+
+
+class AuthSettings(ValueObject):
+    token: TokenVerifierSettings
+    oauth2: OAuth2Settings = OAuth2Settings()
+    scope_verifier: ScopeVerifier = default_scope_verifier
+
+
+def set_auth_scheme(settings: AuthSettings | None) -> OAuth2Schema | None:
+    global scheme
+
+    if settings is not None:
+        scheme = OAuth2Schema(settings)
+
+    return scheme